+++ CHANGES.rst
@@ -5,10 +5,10 @@ Changes to be released in next version
  * 
 
 🙌 Improvements
- * 
+ * DesignKit: Add Fonts (#4356).
 
 🐛 Bugfix
- * 
+ * SSO: Handle login callback URL with HTML entities (#4129).
 
 ⚠️ API Changes
  *